+++ R/aoa.R
@@ -14,7 +14,8 @@
 #' @param weight A data.frame containing weights for each variable. Only required if no model is given.
 #' @param variables character vector of predictor variables. if "all" then all variables
 #' of the model are used or if no model is given then of the train dataset.
-#' @param thres numeric vector of probability of DI in training data, with values in [0,1].
+#' @param thres optional. numeric vector of probability of DI in training data, with values in [0,1] used to define the AOA.
+#' If not specified the DI threshold is 1,5×IQR of the DI of the training data.
 #' @param folds Numeric or character. Folds for cross validation. E.g. Spatial cluster affiliation for each data point.
 #' Should be used if replicates are present. Only required if no model is given.
 #' @param returnTrainDI A logical: should the DI value of the cross-validated training data be returned as a attribute?
@@ -101,7 +102,7 @@ aoa <- function(newdata,
                 train=NULL,
                 weight=NA,
                 variables="all",
-                thres=0.95,
+                thres=NULL,
                 folds=NULL,
                 returnTrainDI=FALSE) {
 
@@ -261,8 +262,11 @@ aoa <- function(newdata,
   TrainDI <- trainDist_min/trainDist_avrgmean
   AOA_train_stats <- quantile(TrainDI,
                               probs = c(0.25,0.5,0.75,0.9,0.95,0.99,1),na.rm = TRUE)
-  thres <- quantile(TrainDI,probs = thres,na.rm=TRUE)
-
+  if(is.null(thres)){
+    thres <- boxplot.stats(TrainDI)$stats[5]
+  }else{
+    thres <- quantile(TrainDI,probs = thres,na.rm=TRUE)
+  }
   #### Create Mask for AOA and return statistics
   if (inherits(out, "RasterLayer")){
     raster::values(out) <- mindist